added data cleaning file
DO NOT RUN.
This file explains my process for cleaning data using December 22-31 as an example. This dataset is not provided due to upload limits on Github, and frankly I wouldn't recommend replicating this process anyways, especially if you're using data from a different transit agency. Instead, I provide a commented file explaining the process I used to clean and validate the data as well as to generate "lateness" coefficients.